Broke up solver.cpp into several smaller files
