+++ tests/unit_tests/restruct_poc/test_rotating_beam.cpp
@@ -6,6 +6,8 @@
 
 #include "src/restruct_poc/beams.hpp"
 #include "src/restruct_poc/solver.hpp"
+#include "src/restruct_poc/InitializeConstraints.hpp"
+#include "src/restruct_poc/Step.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
 
 namespace openturbine::restruct_poc::tests {
