+++ tests/unit_tests/restruct_poc/test_rotor.cpp
@@ -7,6 +7,8 @@
 
 #include "src/restruct_poc/beams.hpp"
 #include "src/restruct_poc/solver.hpp"
+#include "src/restruct_poc/Step.hpp"
+#include "src/restruct_poc/InitializeConstraints.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
 
 #ifdef OTURB_ENABLE_VTK
